+++ applications/appunorganized/appunorganized-plugins/org.csstudio.rap.rtplot/icons/legend.png
