+++ applications/appunorganized/appunorganized-plugins/org.csstudio.rap.rtplot/src/org/csstudio/swt/rtplot/internal/ToggleLegendAction.java
